work on email MS, rethink events

+++ connector/src/main/java/hr/fer/connector/dto/reports/YearPassengerCount.java
@@ -10,5 +10,5 @@ import lombok.NoArgsConstructor;
 public class YearPassengerCount {
 
     private int year;
-    private int passengerCount;
+    private long passengerCount;
 }
